engine/requirements.txt
Pillow==10.3.0
onnxruntime==1.17.0
numpy==1.26.4
pyinstaller==6.6.0
